Code: separate main loop into its own file & add command line args
